cascade delete works, now for multiple file uploads
