+++ frontend/src/store/properties.js
@@ -1,8 +1,8 @@
-
+import { csrfFetch } from './csrf';
 
 // Define action types as constants
 const GET_PROPERTIES = 'properties/setProperties';
-
+const DELETE_PROPERTY = 'properties/deleteProperty';
 // define action creators
 
 //get all properties
@@ -12,6 +12,12 @@ const getProperties = (properties) => ({
 });
 
 
+//Delete a property by ID
+const deleteOneProperty = (property) => ({
+    type: DELETE_PROPERTY,
+    property,
+})
+
 //define thunks
 
 //get all properties
@@ -19,6 +25,16 @@ export const getAllProperties = () => async (dispatch) => {
   const res = await fetch('/api/properties')
   const properties = await res.json();
   dispatch(getProperties(properties))
+};
+
+//delete a property
+export const deleteProperty = (property) => async (dispatch) => {
+    const response = await csrfFetch(`/api/properties/${property}`, {
+        method: 'delete',
+    })
+    if(response.ok) {
+        dispatch(deleteOneProperty(property));
+    };
 };
 
 
@@ -52,6 +68,11 @@ const propertiesReducer = (state = initialState, action) => {
             const newState = {...state, ...action.properties}
             // action.properties.forEach(property => newState[property.id] = property);
             return newState;
+            case DELETE_PROPERTY: {
+                const newState = {...state};
+                delete newState[action.property]
+                return newState;
+            }
             default:
                 return state;
     }
